+++ R/addArchivistHooks.R
@@ -23,12 +23,12 @@
 #' @author 
 #' Przemyslaw Biecek, \email{przemyslaw.biecek@@gmail.com}
 #'  
+#' @template roxlate-references
+#' @template roxlate-contact
 #' @note
 #' One can specify \code{userTags} as in \link{saveToLocalRepo} for artifacts by adding \code{"tags"} attribute.
 #' See note secion about that in \link{saveToLocalRepo}.
 #' 
-#' Bug reports and feature requests can be sent to \href{https://github.com/pbiecek/archivist/issues}{https://github.com/pbiecek/archivist/issues}
-#'  
 #' @examples
 #' 
 #' \dontrun{