+++ R/atrace.R
@@ -15,6 +15,9 @@
 #' @author 
 #' Przemyslaw Biecek, \email{przemyslaw.biecek@@gmail.com}
 #' 
+#' @template roxlate-references
+#' @template roxlate-contact
+#' 
 #' @examples
 #' # read the object from local directory
 #' createLocalRepo("arepo_test", default=TRUE)