+++ R/restoreLibraries.R
@@ -28,9 +28,8 @@
 #' Marcin Kosinski, \email{m.p.kosinski@@gmail.com} \\
 #' Przemyslaw Biecek, \email{przemyslaw.biecek@@gmail.com}
 #' 
-#' @note 
-#' Bug reports and feature requests can be sent to \href{https://github.com/pbiecek/archivist/issues}{https://github.com/pbiecek/archivist/issues}
-#' 
+#' @template roxlate-references
+#' @template roxlate-contact
 #' 
 #' @examples
 #' \dontrun{